+++ features/steps/managers/__init__.py
@@ -9,3 +9,4 @@ from .keys import KeysManager
 from .organization_service import OrganizationServiceManager
 from .organization_service_policy import OrganizationServicePolicyManager
 from .entity_manager import EntityManager
+from .auth_policy import AuthPolicyManager