+++ features/steps/managers/directory.py
@@ -7,14 +7,23 @@ class DirectoryNotCreated(Exception):
     """A Directory was requested but none existed"""
 
 
+class DirectoriesNotRetrieved(Exception):
+    """Directories haven't been retrieved yet"""
+
+
+class PublicKeysNotRetrieved(Exception):
+    """Public keys have not been retrieved yet"""
+
+
 class DirectoryManager(BaseManager):
 
     def __init__(self, organization_factory):
-        self._directories = []
+        self.current_directories = None
+        self.previous_directory = None
         # self._current_directory_entity_list = []  # Look into removing
-        # self._current_sdk_keys = []  # Look into removing
-        # self.previous_directory = None  # Look into removing
-        self._current_public_keys = []
+        self._current_sdk_keys = self.current_sdk_keys = []
+        self.previous_sdk_keys = None
+        self.current_public_keys = None
         self.current_directory = None
 
         BaseManager.__init__(self, organization_factory)
@@ -30,13 +39,43 @@ class DirectoryManager(BaseManager):
         self.previous_directory = getattr(self, "_current_directory", None)
         self._current_directory = value
 
+    @property
+    def current_directories(self):
+        if self._current_directories is None:
+            raise DirectoriesNotRetrieved
+        return self._current_directories
+
+    @current_directories.setter
+    def current_directories(self, value):
+        self.previous_directories = getattr(self, "_current_directories", None)
+        self._current_directories = value
+
+    @property
+    def current_public_keys(self):
+        if self._current_public_keys is None:
+            raise PublicKeysNotRetrieved
+        return self._current_public_keys
+
+    @current_public_keys.setter
+    def current_public_keys(self, value):
+        self.previous_public_keys = getattr(self, "_current_public_keys", None)
+        self._current_public_keys = value
+
+    @property
+    def current_sdk_keys(self):
+        return self._current_sdk_keys
+
+    @current_sdk_keys.setter
+    def current_sdk_keys(self, value):
+        self.previous_sdk_keys = getattr(self, "_current_sdk_keys", None)
+        self._current_sdk_keys = value
+
     def create_directory(self, name=None):
         if name is None:
             name = str(uuid4())
         directory_id = self._organization_client.create_directory(name)
         directory = self._organization_client.get_directory(directory_id)
         self.current_directory = directory
-        self._directories.append(directory)
         return self.current_directory
 
     def get_any_directory(self):
@@ -55,18 +94,31 @@ class DirectoryManager(BaseManager):
         )
 
     def generate_and_add_directory_sdk_key_to_directory(self, directory_id):
-        return self._organization_client.generate_and_add_directory_sdk_key(
+        sdk_key = self._organization_client.generate_and_add_directory_sdk_key(
             directory_id
         )
+        self.current_sdk_keys = self.current_sdk_keys + [sdk_key]
+        return sdk_key
 
     def retrieve_directory(self, directory_id):
-        return self._organization_client.get_directory(directory_id)
+        self.current_directory = self._organization_client.get_directory(directory_id)
+        return self.current_directory
 
     def retrieve_all_directories(self):
-        return self._organization_client.get_all_directories()
+        self.current_directories = \
+            self._organization_client.get_all_directories()
+        return self.current_directories
+
+    def retrieve_directories(self, directory_ids):
+        self.current_directories = \
+            self._organization_client.get_directories(directory_ids)
+        return self.current_directories
 
     def retrieve_directory_sdk_keys(self, directory_id):
-        raise NotImplemented()
+        sdk_keys = self._organization_client.get_all_directory_sdk_keys(
+            directory_id)
+        self.current_sdk_keys = sdk_keys
+        return self.current_sdk_keys
 
     def remove_sdk_key_from_directory(self, sdk_key, directory_id):
         self._organization_client.remove_directory_sdk_key(
@@ -83,17 +135,13 @@ class DirectoryManager(BaseManager):
     def retrieve_directory_public_keys(self, directory_id):
         public_keys = self._organization_client.get_directory_public_keys(
             directory_id)
-        if directory_id == self.current_directory.id:
-            self._current_public_keys = public_keys
+        self.current_public_keys = public_keys
         return public_keys
 
     def update_directory_public_key(self, key_id, directory_id,
                                     expires=False, active=None):
         self._organization_client.update_directory_public_key(
             directory_id, key_id, expires=expires, active=active)
-        if directory_id == self.current_directory.id:
-            # If we update the current directory, trigger a refresh on its keys
-            self.retrieve_directory_public_keys()
 
     def remove_directory_public_key(self, key_id, directory_id):
         self._organization_client.remove_directory_public_key(